fix calculated + consolidate tests locations
