+++ source/dotnet/wholesale-api/CalculationResults/CalculationResults.UnitTests/Infrastructure/SqlStatements/Mappers/AggregationLevelMapperTest.cs
@@ -16,6 +16,7 @@ using Energinet.DataHub.Wholesale.CalculationResults.Infrastructure.SqlStatement
 using Energinet.DataHub.Wholesale.CalculationResults.Infrastructure.SqlStatements.Mappers;
 using Energinet.DataHub.Wholesale.CalculationResults.Interfaces.CalculationResults.Model;
 using FluentAssertions;
+using Test.Core;
 using Xunit;
 using Xunit.Categories;
 
@@ -24,6 +25,32 @@ namespace Energinet.DataHub.Wholesale.CalculationResults.UnitTests.Infrastructur
 [UnitTest]
 public class AggregationLevelMapperTests
 {
+    [Theory]
+    [InlineData(TimeSeriesType.Production, null, null)]
+    [InlineData(TimeSeriesType.Production, "es_id", null)]
+    [InlineData(TimeSeriesType.Production, "es_id", "brp_id")]
+    [InlineData(TimeSeriesType.Production, null, "brp_id")]
+    [InlineData(TimeSeriesType.FlexConsumption, null, null)]
+    [InlineData(TimeSeriesType.FlexConsumption, "es_id", null)]
+    [InlineData(TimeSeriesType.FlexConsumption, "es_id", "brp_id")]
+    [InlineData(TimeSeriesType.FlexConsumption, null, "brp_id")]
+    [InlineData(TimeSeriesType.NonProfiledConsumption, null, null)]
+    [InlineData(TimeSeriesType.NonProfiledConsumption, "es_id", null)]
+    [InlineData(TimeSeriesType.NonProfiledConsumption, "es_id", "brp_id")]
+    [InlineData(TimeSeriesType.NonProfiledConsumption, null, "brp_id")]
+    public async Task ToDeltaTableValue_ReturnsValidDeltaValue(TimeSeriesType timeSeriesType, string? energySupplierId, string? balanceResponsibleId)
+    {
+        // Arrange
+        await using var stream = EmbeddedResources.GetStream("DeltaTableContracts.Contracts.Enums.aggregation-level.json");
+        var validDeltaValues = await ContractComplianceTestHelper.GetCodeListValuesAsync(stream);
+
+        // Act
+        var actual = AggregationLevelMapper.ToDeltaTableValue(timeSeriesType, energySupplierId, balanceResponsibleId);
+
+        // Assert
+        actual.Should().BeOneOf(validDeltaValues);
+    }
+
     [Theory]
     [InlineData(TimeSeriesType.Production)]
     [InlineData(TimeSeriesType.FlexConsumption)]
