+++ source/dotnet/wholesale-api/Events/Events.Infrastructure/IntegrationEvents/calculation_result_completed.proto
@@ -194,6 +194,7 @@ enum QuantityQuality {
   QUANTITY_QUALITY_MEASURED = 2;
   QUANTITY_QUALITY_MISSING = 3;
   QUANTITY_QUALITY_INCOMPLETE = 4;
+  QUANTITY_QUALITY_CALCULATED = 5;
 }
 
 /*
