+++ source/dotnet/wholesale-api/Events/Events.UnitTests/Infrastructure/IntegrationEvents/Mappers/QuantityQualityMapperTests.cs
@@ -27,16 +27,10 @@ public class QuantityQualityMapperTests
     [InlineAutoMoqData(QuantityQuality.Incomplete, Contracts.Events.QuantityQuality.Incomplete)]
     [InlineAutoMoqData(QuantityQuality.Measured, Contracts.Events.QuantityQuality.Measured)]
     [InlineAutoMoqData(QuantityQuality.Missing, Contracts.Events.QuantityQuality.Missing)]
+    [InlineAutoMoqData(QuantityQuality.Calculated, Contracts.Events.QuantityQuality.Calculated)]
     public void MapQuantityQuality_WhenCalled_MapsCorrectly(QuantityQuality quantityQuality, Wholesale.Contracts.Events.QuantityQuality expected)
     {
         // Act & Assert
         QuantityQualityMapper.MapQuantityQuality(quantityQuality).Should().Be(expected);
-    }
-
-    [Fact]
-    public void MapQuantityQuality_WhenCalledWithCalculated_ExceptionIsThrown()
-    {
-        // Act & Assert
-        Assert.Throws<ArgumentException>(() => QuantityQualityMapper.MapQuantityQuality(QuantityQuality.Calculated));
     }
 }
